refactor: Create tag/tags filter class (#714)
* Extract variables in Query.parseTagFilter()

Preparing to move method to new TagsField class

* Move the tag and tags search code to TagsField

* Move filtering test functions to FilterTestHelpers.ts

* Move tag tests from Query.test.ts to TagsField.test.ts

They are still based on Query for now, so test TagsField indirectly.
